requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
Flask==3.1.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.4
MarkupSafe==3.0.2
protobuf==5.29.4
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
redis==6.0.0
setuptools==80.3.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
tenacity>=8.2.2
pybreaker>=0.7.0
pydantic-settings>=0.1.1
transformers==4.51.3
torch==2.7.0
